Switch to new command format
These commands are now send inside Embeds, which means inline/named links
are now possible. In case this feature is not wanted by the user, all commands
with in-line links also have an optional parameter to turn them off.

Under the hood, commands are no longer read from the filesystem everytime they're
ran (resulting in less drive usage, but requiring a full bot restart to reload
commands), and are instead read in at startup and cached.

Command attachments are now read from the 'attachments' folder, and are sent as a
seperate message (since sending them with the Embed would result in them showing up
before the Embed, which looks weird).

Commands also support a 'raw_message' attribute, which sends the text defined as a
regular message, in case any formatting errors ever arise (see 'ramplacement.yml' for
an example of that)