Add 'up_to_date_range'
This allows to specify more than one Windows version to be "Up to date".
This is intended for version transitions (like right now), where (as an
example) 21H2 got released, but only some people are getting the update
and it might have issues.

The variable in the config file is the maximum version difference that's
still considered up-to-date (1 -> the latest version + 1 more
version are considered up-to-date)